+++ src/Vankosoft/ApplicationBundle/Resources/themes/default/assets/js/pages/sliders-edit.js
@@ -1,7 +1,9 @@
 require( '../includes/resource-delete.js' );
 require( '../includes/bootstrap-5/file-input.js' );
 
+// bin/console fos:js-routing:dump --format=json --target=public/shared_assets/js/fos_js_routes_admin.json
 import { VsPath } from '../includes/fos_js_routes.js';
+
 import { VsTranslator, VsLoadTranslations } from '../includes/bazinga_js_translations.js';
 VsLoadTranslations(['VSCmsBundle']);